Turn `mockDataToSubgraph` into a hook
