+++ packages/mock-block-dock/src/datastore/use-mock-data-to-subgraph.ts
@@ -1,0 +1,25 @@
+import { Subgraph } from "@blockprotocol/graph";
+import { useMemo } from "react";
+
+import { addEntitiesToSubgraph } from "./mutate-subgraph";
+import { MockData } from "./use-mock-datastore";
+
+export const useMockDataToSubgraph = (mockData: MockData): Subgraph => {
+  return useMemo(() => {
+    const { entities } = mockData;
+
+    const subgraph: Subgraph = {
+      roots: [],
+      vertices: {},
+      edges: {},
+      depths: {
+        hasLeftEntity: { incoming: 255, outgoing: 255 },
+        hasRightEntity: { incoming: 255, outgoing: 255 },
+      },
+    };
+
+    addEntitiesToSubgraph(subgraph, entities);
+
+    return subgraph;
+  }, [mockData]);
+};
